fixed bags and refactor
