+++ src/main/java/com/massonus/rccnavigator/controller/UserController.java
@@ -1,4 +1,4 @@
-package com.massonus.rccnavigator.controllers;
+package com.massonus.rccnavigator.controller;
 
 import com.massonus.rccnavigator.entity.Role;
 import com.massonus.rccnavigator.entity.User;
