+++ src/main/java/com/massonus/rccnavigator/entity/User.java
@@ -38,7 +38,7 @@ public class User implements UserDetails {
     @OneToMany(mappedBy = "author", cascade = CascadeType.ALL)
     private List<Message> messages = new ArrayList<>();
 
-    @OneToMany(mappedBy = "author")
+    @OneToMany(mappedBy = "author", cascade = CascadeType.ALL)
     private List<Rating> rates = new ArrayList<>();
 
     private Long redactor;
